Restructured project template file
