+++ src/muvis_align/ui/project_template.yaml
